+++ src/Normalizer/SimpleNormalizer.php
@@ -8,12 +8,16 @@ use Symfony\Component\DependencyInjection\ServiceLocator;
 use Torr\SimpleNormalizer\Exception\ObjectTypeNotSupportedException;
 use Torr\SimpleNormalizer\Exception\UnsupportedTypeException;
 
-final class SimpleNormalizer
+/**
+ * @final
+ */
+readonly class SimpleNormalizer
 {
 	/**
+	 * @param ServiceLocator<SimpleObjectNormalizerInterface> $objectNormalizers
 	 */
 	public function __construct (
-		private readonly ServiceLocator $objectNormalizers,
+		private ServiceLocator $objectNormalizers,
 	) {}
 
 	/**